Move to another npm org scope
The `pl-beta` tag was a placeholder (meaning "psychollama-beta"). I have
other packages I'd like to publish and I don't want to commit to the old
tag.

I created `@blabbing` because it's fun to say, is nonsensical enough
that it works with most words, and loosely aligns with the goal of the
org scope: a place to publish my private packages for testing (aka
"blabbing" my private projects).