Automate changelog and release process
Now I have a manual workflow that automatically updates the changelog,
creates a git commit/tag with release notes, and pushes a tag (which
triggers the stable release flow).

All I have to do is enter a version and press a button.

This is overkill for the current project, but I'm using it as a test
before implementing it for my other projects.