Return promises from actions
This is a big change, both code-wise and philosophically, but I'm
convinced it's the right one. It replaces `suspendEffects` by returning
the `Promise.all` from the executed action.

This has a number of advantages. First, if you return an ID or atom
reference from an action the caller can wait for the behavior to
initialize it before using it. Second, it makes unit tests and error
handling more conventional because errors are no longer implicitly
swallowed. Third, it makes it clear that actions can have consequences
(er, behaviors).

I'm happy with this solution.

The caveat is you can no longer nest actions. The promise chain got
weird. I think I could've done it safely, but IMO it would be a bad idea
to try. It's not worth the cost.